RR-173 Added tests around trade agreements Refactored the validators
