+++ test/selectors/event/create.js
@@ -41,7 +41,7 @@ module.exports = {
   saveEvent: 'button:contains("Add event")',
   saveAndReturn: 'button:contains("Save and return")',
   relatedTradeAgreements: '[name="has_related_trade_agreements"]',
-  addAnotherTradeAgreement: 'input[name="add_related_trade_agreement"]',
+  addAnotherTradeAgreement: 'input[id="add_related_trade_agreement"]',
   tradeAgreementExistsYes: 'label[for="field-has_related_trade_agreements-1"]',
   tradeAgreementError:
     '#group-field-has_related_trade_agreements .c-form-group__error-message',
